Moved Matmodule and added 1d support
